Improved subject key identifiers.
This change updates the import command to generate subject key identifiers
where none are available, and updates the certificate database to add the
missing roots from ca-bundle.crt that were affected by this.

The current certificate database doesn't have the full set of CA
certificates contained in ca-bundle.crt. This is due to the way that the
database is set up: namely, a unique constraint on the tuple of (subject
key identifier, serial number). The SKI is taken directly from the
*x509.Certificate structure, which is taken directly from the field in
the certificate.

The issue here is that there are approximately 15 root certificates (it
looks like they were all issued between 1996 and 2000) that do not have
this field in the certificate, resulting in an empty SKI. They also all
have the serial number of 0 or 1. This means that the insertion logic
fails to add them into the database; it thinks they're already present.

The workaround here is to give these certificates an SKI using the algorithm
specified in RFC 5280, section 4.2.1.2 (1). Note that these algorithms are a
*suggestion*, and that it's possible these roots are issuing certificates
using an SKI different from what this change uses. In the absence of having
any certificates signed by these roots on hand to verify, this seems like a
sane default (though "sane" and "certificate authority" ought not to be in
the same sentence).

It's possible that this isn't the right approach, and that we may have to
change this in the future. It does, however, provide the same set of roots
that this repository currently has.