Update the trust bundles.
This change uses the cfssl-trust utility to produce the bundles. This
has the effect of de-duplicating the certificate bundles, which reduces
the size of the CA bundle by about 75% and the intermediate bundle by
25%.

As evidence that the new bundles are equivalent to the previous bundles,
see https://gist.github.com/kisom/f097957f64d3c5d8c7c491b29cbb2011.